rename subpackage 'hibernate' to more general 'model'
