+++ src/main/java/de/fuberlin/chaostesting/UserTest.java
@@ -2,7 +2,7 @@ package de.fuberlin.chaostesting;
 
 import java.util.List;
 
-import de.fuberlin.chaostesting.hibernate.User;
+import de.fuberlin.chaostesting.model.User;
 
 public class UserTest {
 		public String newTester() {
